consolidate back down to less separate commands by using flags
